pyproject.toml
[tool.poetry]
name = "jst-django-core"
version = "v1.1.9"
description = ""
authors = ["A'zamov Samandar <JscorpTech@gmail.com>"]
readme = "README.md"
packages = [{ include = "django_core", from = "src" }]

[tool.poetry.dependencies]
python = "*"
django = "*"
djangorestframework = "*"
celery = "*"
django-unfold = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
